refactor(cpuhealth): one file per source for cpu.stat and cpu.max
Give the cgroup's cpu.stat and cpu.max readers their own files, matching
every other CPU source (/proc/stat, cpu.pressure, cpuset.cpus.effective),
each of which already lives in its own file next to its own reader.
